requirements.txt
dash
plotly
numpy
Pillow
cellpose
